+++ src/options/presets/Default.ts
@@ -44,8 +44,8 @@ export const DEFAULT_PRESET: TInputOptions = Object.freeze({
     stringArrayEncoding: [
         StringArrayEncoding.None
     ],
-    stringArrayWrappersChainedCalls: false,
-    stringArrayWrappersCount: 0,
+    stringArrayWrappersChainedCalls: true,
+    stringArrayWrappersCount: 1,
     stringArrayThreshold: 0.75,
     target: ObfuscationTarget.Browser,
     transformObjectKeys: false,