+++ src/options/presets/MediumObfuscation.ts
@@ -16,7 +16,6 @@ export const MEDIUM_OBFUSCATION_PRESET: TInputOptions = Object.freeze({
     stringArrayEncoding: [
         StringArrayEncoding.Base64
     ],
-    stringArrayWrappersChainedCalls: true,
     stringArrayWrappersCount: 2,
     transformObjectKeys: true
 });